+++ lib/moon/design/bottom_sheet/backdrop.ex
@@ -3,8 +3,13 @@ defmodule Moon.Design.BottomSheet.Backdrop do
 
   use Moon.StatelessComponent, slot: "backdrop"
 
+  @doc "Tailwind classes for customization"
   prop(class, :css_class)
+
+  @doc "Data-testid attribute for DOM element"
   prop(testid, :string)
+
+  @doc "Id attribute for DOM element"
   prop(id, :string)
 
   defdelegate render(assigns), to: Moon.Design.Modal.Backdrop