+++ lib/moon/design/bottom_sheet/panel.ex
@@ -3,17 +3,29 @@ defmodule Moon.Design.BottomSheet.Panel do
 
   use Moon.StatelessComponent, slot: "panel"
 
-  slot(draghandle)
-  slot(default)
-
+  @doc "Tailwind classes for customization"
   prop(class, :css_class)
 
+  @doc "Data-testid attribute for DOM element"
   prop(testid, :string)
+
+  @doc "Id attribute for DOM element"
   prop(id, :string)
 
+  @doc "Header of BottomSheet, see BottomSheet.Header"
+  slot(header)
+
+  @doc "Content of BottomSheet"
+  slot(default)
+
   # inner API
+  @doc "Will be got from Bottomsheet in most cases"
   prop(on_close, :event, from_context: :on_close)
+
+  @doc "Will be got from Bottomsheet in most cases"
   prop(has_shadow, :boolean, from_context: :has_shadow)
+
+  @doc "Will be got from Bottomsheet in most cases"
   prop(size, :string, from_context: :size)
 
   def render(assigns) do
@@ -24,7 +36,7 @@ defmodule Moon.Design.BottomSheet.Panel do
       data-animate_leave_class="animate-drawer_leave_bottom"
       class={merge([
         [
-          "fixed w-full bottom-0 rounded-t-xl bg-gohan",
+          "fixed w-full bottom-0 rounded-t-xl bg-gohan overflow-y-auto",
           "moon-panel",
           "shadow-moon-lg": @has_shadow,
           "h-[32%]": @size == "sm",
@@ -36,7 +48,7 @@ defmodule Moon.Design.BottomSheet.Panel do
       :on-click-away={@on_close}
       data-testid={@testid}
     >
-      <#slot {@draghandle} />
+      <#slot {@header} />
       <#slot />
     </div>
     """