+++ lib/moon/design/form/dropdown.ex
@@ -1,5 +1,5 @@
 defmodule Moon.Design.Form.Dropdown do
-  @moduledoc "Totally styled dropdown component for the forms"
+  @moduledoc "Fully styled select component for the forms"
 
   use Moon.StatelessComponent
 
@@ -9,16 +9,13 @@ defmodule Moon.Design.Form.Dropdown do
   alias Moon.Design.Form.Checkbox
   alias Moon.Design.Form.Radio
 
-  import Moon.Helpers.Form, only: [has_error: 2]
-  import Phoenix.HTML.Form, only: [input_id: 2, input_value: 2]
-
   @doc "Name of the field, usually should be taken from context"
   prop(field, :atom, from_context: {Form.Field, :field})
   @doc "Form info, usually should be taken from context"
   prop(form, :form, from_context: {Form, :form})
 
   @doc "... format: [%{key: shown_label, value: option_value, disabled: bool}], diisabled is optional"
-  prop(options, :generator, required: true)
+  prop(options, :list, required: true)
 
   @doc "Selected option(s) value - do not use it inside the form, just for away-from-form components"
   prop(value, :any)
@@ -30,72 +27,85 @@ defmodule Moon.Design.Form.Dropdown do
   @doc "Additional classes for the <select> tag"
   prop(class, :css_class, from_context: :class)
   @doc "Some prompt to be shown on empty value"
-  prop(prompt, :string)
+  prop(prompt, :string, default: "...")
 
   @doc "Id to be given to the select tag"
   prop(id, :string)
   @doc "Data-testid attribute value"
   prop(testid, :string)
 
-  @doc "Some additional styling will be set to indicate field is iinvalid. generally should be set by Form.Field component"
+  @doc "Some additional styling will be set to indicate field is iinvalid"
   prop(error, :boolean, from_context: :error)
 
   @doc "If field does support multiselect, `multiple` attribute for select tag in HTML terms"
   prop(is_multiple, :boolean)
 
-  @doc "Option for custom stylings - use it to put icons or anything else"
-  slot(option, generator_prop: :options)
+  @doc "Should dropdown be open"
+  prop(is_open, :boolean)
+
+  @doc "Option for custom stylings - use it to show icons or anything else"
+  slot(default)
+
+  @doc "Trigger element for the dropdown, default is Dropdown.Select"
+  slot(trigger)
+
+  @doc "Slot used for rendering single option. option[:key] will be used if not given"
+  slot(option)
 
   def render(assigns) do
     ~F"""
-    <Dropdown id={"#{@id || input_id(@form, @field)}-dropdown"}>
-      <Dropdown.Trigger>
-        <input
-          class={merge([
-            "focus:ring-0 border-0 py-0 px-4 m-0",
-            "block w-full max-w-full appearance-none text-[1rem] text-bulma transition-shadow box-border before:box-border after:box-border",
-            "relative z-[2] shadow-input hover:shadow-input-hov focus:shadow-input-focus focus:outline-none bg-gohan h-10",
-            "placeholder:text-trunks placeholder:opacity-100 placeholder:transition-opacity placeholder:delay-75",
-            "read-only:outline-0 read-only:border-none read-only:cursor-not-allowed read-only:hover:shadow-input read-only:focus:shadow-input",
-            "moon-error:shadow-input-err moon-error:hover:shadow-input-err moon-error:focus:shadow-input-err",
-            "invalid:shadow-input-err invalid:hover:shadow-input-err invalid:focus:shadow-input-err",
-            [
-              "leading-8 rounded-moon-i-xs": @size == "sm",
-              "leading-10 rounded-moon-i-sm": @size == "md",
-              "leading-[3rem] rounded-moon-i-sm": @size == "lg",
-              "opacity-30": @disabled
-            ],
-            @class
-          ])}
-          type="text"
-          placeholder={@prompt}
-          value={input_value(@form, @field)}
-          {=@disabled}
-          error={@error || has_error(@form, @field)}
-          data-testid={@testid}
-        />
-      </Dropdown.Trigger>
-      <Dropdown.Options>
-        <#slot {@option} :for={option <- @options} generator_value={option}>
-          <Dropdown.Option value={option[:value]} disabled={option[:disabled]}>
+    <Dropdown id={dropdown_id(assigns)} {=@is_open}>
+      <:trigger :let={is_open: is_open}>
+        <#slot {@trigger, is_open: is_open}>
+          <Dropdown.Select
+            {=@prompt}
+            {=@size}
+            {=is_open}
+            {=@error}
+            {=@disabled}
+            value={select_value(assigns)[:key]}
+            badge={select_badge(assigns)}
+          ><#slot {@option, option: select_value(assigns)}>{select_value(assigns)[:key] || @prompt}</#slot></Dropdown.Select>
+        </#slot>
+      </:trigger>
+      <#slot {@default}>
+        <Dropdown.Options on_change={option_click()}>
+          <Dropdown.Option :for={option <- @options} {=@size}>
             <Checkbox
               checked_value={option[:value]}
               :if={@is_multiple}
-              label={option[:key]}
               disabled={option[:disabled]}
+              {=@size}
               hidden_input={false}
               is_multiple
-            />
-            <Radio.Button
-              value={option[:value]}
-              :if={!@is_multiple}
-              label={option[:key]}
-              disabled={option[:disabled]}
-            />
+            >
+              <#slot {@option, option: option}>{option[:key]}</#slot>
+            </Checkbox>
+            <Radio.Button value={option[:value]} :if={!@is_multiple} disabled={option[:disabled]} {=@size}>
+              <Radio.Indicator />
+              <#slot {@option, option: option}>{option[:key]}</#slot>
+            </Radio.Button>
           </Dropdown.Option>
-        </#slot>
-      </Dropdown.Options>
+        </Dropdown.Options>
+      </#slot>
     </Dropdown>
     """
   end
+
+  defp dropdown_id(%{form: form, field: field, id: id}), do: "#{id || form[field].id}-dropdown"
+
+  defp option_click(), do: Phoenix.LiveView.JS.dispatch("moon-empty-event")
+
+  defp select_value(%{is_multiple: true}), do: nil
+
+  defp select_value(%{form: form, field: field, options: options}) do
+    options |> Enum.find(&(&1[:value] == form[field].value))
+  end
+
+  defp select_badge(%{is_multiple: true, form: form, field: field}) do
+    (form[field].value && Enum.count(form[field].value) > 0 && Enum.count(form[field].value)) ||
+      nil
+  end
+
+  defp select_badge(_), do: nil
 end