+++ lib/moon_web/examples/design/form/input_example/with_label.ex
@@ -8,7 +8,7 @@ defmodule MoonWeb.Examples.Design.Form.InputExample.WithLabel do
   alias Moon.Design.Form
   alias Moon.Design.Form.Input
   alias Moon.Design.Form.Field
-  alias MoonWeb.Pages.Tutorials.AddDataUsingForm.User
+  alias MoonWeb.Schema.User
 
   prop(user_changeset, :any, default: User.changeset(%User{}))
 
@@ -35,7 +35,7 @@ defmodule MoonWeb.Examples.Design.Form.InputExample.WithLabel do
     alias Moon.Design.Form
     alias Moon.Design.Form.Input
     alias Moon.Design.Form.Field
-    alias MoonWeb.Pages.Tutorials.AddDataUsingForm.User
+    alias MoonWeb.Schema.User
 
     prop(user_changeset, :any, default: User.changeset(%User{}))