+++ lib/moon_web/examples/design/form/select_example/default.ex
@@ -5,7 +5,7 @@ defmodule MoonWeb.Examples.Design.Form.SelectExample.Default do
 
   use MoonWeb, :example
 
-  alias MoonWeb.Pages.Tutorials.AddDataUsingForm.User
+  alias MoonWeb.Schema.User
   alias Moon.Design.Form
   alias Moon.Design.Form.Select
 
@@ -30,7 +30,7 @@ defmodule MoonWeb.Examples.Design.Form.SelectExample.Default do
 
   def code() do
     """
-    alias MoonWeb.Pages.Tutorials.AddDataUsingForm.User
+    alias MoonWeb.Schema.User
     alias Moon.Design.Form
     alias Moon.Design.Form.Select